Use GET instead of HEAD request to check if the server is ready
When we use HEAD request, 'start-server-and-test' assumes that the server
is ready after the server responses.
However, at that point of time, the Gatsby doesn't necessarily ready to serve
the page content. If we start testing at that point of time, Cypress
won't be able to load the page. Hence the early tests are failing.
By using GET request instead, we only start the test after Gatsby is
confirmed ready to serve the pages, so we can ensure that the early
tests are always successful.

For more information on the 'start-server-and-test' API, please refer to this section:
https://github.com/bahmutov/start-server-and-test#note-for-webpack-dev-server-users